bullpholio/requirements.txt
pymupdf>=1.24.0
pdfplumber>=0.11.0

pandas>=2.1.0
openpyxl>=3.1.2

python-docx>=1.1.0

pillow>=10.0.0
opencv-python-headless>=4.9.0
easyocr>=1.7.1

numpy>=1.26.0

pydantic>=2.6.0
python-dotenv>=1.0.1

scikit-image>=0.22.0
matplotlib>=3.8.0